fix(parametermanager): update test cases and move samples to api folder
