+++ parametermanager/api/ParameterManager.Samples.Tests/CreateStructuredParameterVersionTests.cs
@@ -31,12 +31,16 @@ public class CreateStructureParameterVersionTests
     [Fact]
     public void CreateStructureParameterVersion()
     {
-        ParameterVersionName parameterVersionName = _fixture.ParameterVersionNameWithFormat;
-        string payload = ParameterManagerFixture.JsonPayload;
+        ParameterVersionName parameterVersionName = new ParameterVersionName(_fixture.ProjectId, ParameterManagerFixture.LocationId, _fixture.RandomId(), _fixture.RandomId());
+        Parameter parameter = _fixture.CreateParameter(parameterVersionName.ParameterId, ParameterFormat.Json);
+        string payload = "{\"username\": \"test-user\", \"host\": \"localhost\"}";
         ParameterVersion result = _sample.CreateStructuredParameterVersion(
           projectId: parameterVersionName.ProjectId, parameterId: parameterVersionName.ParameterId, versionId: parameterVersionName.ParameterVersionId, payload: payload);
 
         Assert.NotNull(result);
         Assert.Equal(result.ParameterVersionName.ParameterVersionId, parameterVersionName.ParameterVersionId);
+
+        _fixture.ParametersToDelete.Add(parameter.ParameterName);
+        _fixture.ParameterVersionsToDelete.Add(parameterVersionName);
     }
 }
