+++ parametermanager/api/ParameterManager.Samples.Tests/ParameterManager.Samples.Tests.csproj
