+++ parametermanager/api/ParameterManager.Samples/CreateParameterVersionWithSecret.cs
