+++ parametermanager/api/ParameterManager.Samples/ParameterManager.Samples.csproj
